Bar Code printing on the patient creation confirmation page.
I am decently sure that this works, unfortunately, I cannot fully test it because I had to give the printer back to Rich.  When the print bar code button is clicked, it sends a document to the printer.  It will print the same label as if you clicked the print label button on the patient details page.